+++ AzzyBot-Next/Commands/Autocompletes/GuildsAutocomplete.cs
@@ -26,12 +26,12 @@ public sealed class GuildsAutocomplete(DbActions dbActions, DiscordBotService bo
 
         switch (context.Command.FullName)
         {
-            case "admin remove-debug-guild":
-                guildsInDb = await _dbActions.GetGuildsWithDebugAsync(true);
+            case "admin debug-servers add-server":
+                guildsInDb = await _dbActions.GetGuildsWithDebugAsync(false);
                 break;
 
-            case "admin set-debug-guild":
-                guildsInDb = await _dbActions.GetGuildsWithDebugAsync(false);
+            case "admin debug-server remove-server":
+                guildsInDb = await _dbActions.GetGuildsWithDebugAsync(true);
                 break;
         }